refactor: refactoring models
JupiterModel -> JupiterPostInputModel, OutputModel, ServiceModel
