+++ TJLabsJupiter/Model/JupiterServiceModel.swift
@@ -1,0 +1,78 @@
+
+import Foundation
+import TJLabsCommon
+
+public enum JupiterRegion {
+    case KOREA, US, CANADA
+}
+
+
+// MARK: - FineLocationTracking
+public struct TrajectoryInfo {
+    public var index: Int = 0
+    public var length: Double = 0
+    public var heading: Double = 0
+    public var velocity: Double = 0
+    public var lookingFlag: Bool = false
+    public var isIndexChanged: Bool = false
+    public var numBleChannels: Int = 0
+    public var scc: Double = 0
+    public var userBuilding: String = ""
+    public var userLevel: String = ""
+    public var userX: Double = 0
+    public var userY: Double = 0
+    public var userHeading: Double = 0
+    public var userPmSuccess: Bool = false
+    public var userTuHeading: Double = 0
+}
+
+public struct SearchInfo {
+    public var searchRange: [Int] = []
+    public var searchDirection: [Int] = [0, 90, 180, 270]
+    public var tailIndex: Int = 1
+}
+
+// MARK: - JupiterResult
+public struct JupiterResult: Codable {
+    public var mobile_time: Int
+    public var building_name: String
+    public var level_name: String
+    public var scc: Double
+    public var x: Double
+    public var y: Double
+    public var absolute_heading: Double
+    public var phase: Int
+    public var calculated_time: Double
+    public var index: Int
+    public var velocity: Double
+    public var mode: UserMode
+    public var ble_only_position: Bool
+    public var isIndoor: Bool
+    public var validity: Bool
+    public var validity_flag: Int
+    
+//    public init() {
+//        self.mobile_time = 0
+//        self.building_name = ""
+//        self.level_name = ""
+//        self.scc = 0
+//        self.x = 0
+//        self.y = 0
+//        self.absolute_heading = 0
+//        self.phase = 0
+//        self.calculated_time = 0
+//        self.index = 0
+//        self.velocity = 0
+//        self.mode = .MODE_PEDESTRIAN
+//        self.ble_only_position = false
+//        self.isIndoor = false
+//        self.validity = false
+//        self.validity_flag = 0
+//    }
+}
+
+public protocol JupiterManagerDelegate: AnyObject {
+    func onJupiterSuccess(_ isSuccess: Bool)
+    func onJupiterError(_ code: Int, _ msg: String)
+    func onJupiterResult(_ result: JupiterResult)
+}
